requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.8.1
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
cppimport==21.3.7
discord.py==1.7.3
idna==3.3
Mako==1.1.6
MarkupSafe==2.1.0
multidict==6.0.2
mysql-connector-python==8.0.28
Pillow==9.0.1
protobuf==3.19.4
pybind11==2.9.1
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urbandictionary==1.1
urllib3==1.26.8
yarl==1.7.2
